Why include an extra trendline?
This is not a PR I feel strongly about, but looking over the lab, the code we are providing is making a regression line and a trendline based on loess. I wasn't able to intuit if there was some lesson that students were supposed to get by looking at the difference between the two. Is there? 

The text of the lab seems to not refer to the trendline directly. I would argue unless there is some insight I am missing, the trendline isn't adding much. 

Looking at the auto-generated plots for the sensitivity analysis, my gut instinct was to try to find the differences between the two lines on the plot to answer the questions --even though it seems like they are telling the same story.